// src/app/components/admin-store/order/calendar/view_calendar/types.ts
export interface Order {
  id: any
  title: string
  start: string
  end: string
  status_order?: string
  order_status?: string
  className: string
  order_detail?: any
}

export interface OrderHistory {
  order_id: number
  order_status: string
  created_at: string
  updated_by: string
}

export interface Status {
  value: number | null
  category: string
}

export interface PaymentStage {
  stage: string
  percentage: string
  amount: number
}